+++ src/capabilities/confscan.h
@@ -35,12 +35,13 @@
 #include "curcumamethod.h"
 
 constexpr double third = 1 / 3.0;
-/*
+#ifdef WriteMoreInfo
 struct dnn_input {
     double dE, dIa, dIb, dIc, dH, rmsd;
     Matrix dHM;
 };
-*/
+#endif
+
 static const json ConfScanJson = {
     { "noname", true },
     { "restart", true },
@@ -55,12 +56,12 @@ static const json ConfScanJson = {
     { "preventreorder", false },
     { "scaleLoose", 1.5 },
     { "scaleTight", 0.1 },
-    { "scaleLooseEnergy", 1.2 },
-    { "scaleTightEnergy", 0.1 },
-    { "scaleLooseRotational", 1.2 },
-    { "scaleTightRotational", 0.1 },
-    { "scaleLooseRipser", 1.2 },
-    { "scaleTightRipser", 0.1 },
+    { "sLE", 1.2 },
+    { "sTE", 0.1 },
+    { "sLI", 1.2 },
+    { "sTI", 0.1 },
+    { "sLH", 1.2 },
+    { "sTH", 0.1 },
     { "skip", 0 },
     { "allxyz", false },
     { "update", false },
@@ -97,7 +98,8 @@ static const json ConfScanJson = {
     { "ripser_stdy", 10 },
     { "ripser_ratio", 1 },
     { "ripser_dimension", 2 },
-    { "domolalign", -1 }
+    { "domolalign", -1 },
+    { "molaligntol", 10 }
 };
 
 class ConfScanThread : public CxxThread {
@@ -133,6 +135,9 @@ public:
     void setTarget(const Molecule* molecule)
     {
         m_target.setGeometry(molecule->getGeometry());
+        m_target.setPersisentImage(molecule->getPersisentImage());
+        m_target.CalculateRotationalConstants();
+        m_target.setEnergy(molecule->Energy());
     }
     std::vector<int> ReorderRule() const { return m_reorder_rule; }
     void setReorderRules(const std::vector<std::vector<int>>& reorder_rules)
@@ -147,22 +152,27 @@ public:
 
     double RMSD() const { return m_rmsd; }
     const Molecule* Reference() const { return &m_reference; }
+    double Energy() const { return m_energy; }
+#ifdef WriteMoreInfo
     void setPredRMSD(double rmsd) { m_pred_rmsd = rmsd; }
     double PredRMSD() const { return m_pred_rmsd; }
-    // dnn_input getDNNInput() const { return m_input; }
+    dnn_input getDNNInput() const { return m_input; }
+#endif
 
 private:
     bool m_keep_molecule = true, m_break_pool = false, m_reorder_worked = false, m_reuse_only = false, m_reused_worked = false;
     Molecule m_reference, m_target;
-    double m_rmsd = 0, m_rmsd_threshold = 1;
+    double m_rmsd = 0, m_rmsd_threshold = 1, m_energy = 0;
     int m_MaxHTopoDiff;
     int m_threads = 1;
     std::vector<int> m_reorder_rule;
     std::vector<std::vector<int>> m_reorder_rules;
     RMSDDriver* m_driver;
     json m_config;
+#ifdef WriteMoreInfo
     double m_pred_rmsd = 0;
-    // dnn_input m_input;
+    dnn_input m_input;
+#endif
 };
 
 class ConfScanThreadNoReorder : public CxxThread {
@@ -182,8 +192,8 @@ public:
     }
 
     virtual int execute() override;
-    double DiffRot() const { return m_diff_rotational; }
-    double DiffRipser() const { return m_diff_ripser; }
+    double DI() const { return m_DI; }
+    double DH() const { return m_DH; }
     double RMSD() const { return m_rmsd; }
     const Molecule* Reference() const { return &m_reference; }
 
@@ -200,21 +210,29 @@ public:
         m_target.setGeometry(molecule->getGeometry());
         m_target.setPersisentImage(molecule->getPersisentImage());
         m_target.CalculateRotationalConstants();
+        m_target.setEnergy(molecule->Energy());
     }
 
     bool KeepMolecule() const { return m_keep_molecule; }
-    // dnn_input getDNNInput() const { return m_input; }
+#ifdef WriteMoreInfo
+    dnn_input getDNNInput() const
+    {
+        return m_input;
+    }
+#endif
 
 private:
     bool m_keep_molecule = true, m_break_pool = false;
-    double m_diff_rotational = 0, m_diff_ripser = 0;
+    double m_DI = 0, m_DH = 0;
     Molecule m_reference, m_target;
 
     RMSDDriver* m_driver;
     json m_config;
     double m_rmsd = 0, m_rmsd_threshold = 1;
     int m_MaxHTopoDiff;
-    // dnn_input m_input;
+#ifdef WriteMoreInfo
+    dnn_input m_input;
+#endif
 };
 
 class ConfScan : public CurcumaMethod {
@@ -285,7 +303,7 @@ private:
 
     std::vector<std::vector<int>> m_reorder_rules;
 
-    void PrintStatus();
+    void PrintStatus(const std::string& info = "");
 
     std::map<std::string, std::vector<std::string>> m_filtered;
     bool m_ok;
@@ -297,25 +315,28 @@ private:
     std::map<double, int> m_ordered_list;
     std::vector<std::pair<std::string, Molecule*>> m_molecules;
     double m_rmsd_threshold = 1.0, m_nearly_missed = 0.8, m_energy_cutoff = -1, m_reference_last_energy = 0, m_target_last_energy = 0, m_lowest_energy = 1, m_current_energy = 0;
-    double m_scaleTightEnergy = 0.1, m_scaleLooseEnergy = 1.5;
-    double m_scaleTightRotational = 0.1, m_scaleLooseRotational = 1.5;
-    double m_scaleTightRipser = 0.1, m_scaleLooseRipser = 1.5;
+    double m_sTE = 0.1, m_sLE = 1.5;
+    double m_sTI = 0.1, m_sLI = 1.5;
+    double m_sTH = 0.1, m_sLH = 1.5;
 
     double m_reference_restored_energy = -1e10, m_target_restored_energy = -1e10;
-    double m_diff_rot_threshold_loose = 0.0, m_diff_ripser_threshold_loose = 0.0, m_diff_energy_threshold_loose = 0.0;
-    double m_diff_rot_threshold_tight = 0.0, m_diff_ripser_threshold_tight = 0.0, m_diff_energy_threshold_tight = 0.0;
+    double m_dLI = 0.0, m_dLH = 0.0, m_dLE = 0.0;
+    double m_dTI = 0.0, m_dTH = 0.0, m_dTE = 0.0;
 
     std::vector<Molecule*> m_result, m_rejected_structures, m_stored_structures, m_previously_accepted;
     std::vector<const Molecule*> m_threshold;
     std::vector<int> m_element_templates;
-    // std::vector<dnn_input> m_dnn_data;
+
+#ifdef WriteMoreInfo
+    std::vector<dnn_input> m_dnn_data;
+#endif
 
     std::string m_rmsd_element_templates;
     std::string m_method = "";
     std::string m_molalign = "molalign";
 
     double m_domolalign = -1;
-    double m_last_diff = 0.0, m_last_ripser = 0.0, m_last_dE = -1, m_dE = -1, m_damping = 0.8;
+    double m_lastDI = 0.0, m_lastDH = 0.0, m_lastdE = -1, m_dE = -1, m_damping = 0.8;
     int m_maxmol = 0;
     int m_maxrank = 10000;
     int m_maxParam = -1;
@@ -325,6 +346,8 @@ private:
     int m_MaxHTopoDiff = -1;
     int m_threads = 1;
     int m_RMSDElement = 7;
+    int m_molaligntol = 10;
+
     bool m_writeXYZ = false;
     bool m_check_connections = false;
     bool m_force_reorder = false, m_prevent_reorder = false;